fix: group overview API calls
