+++ pages/api/v1/metrics/get-overview.ts
@@ -1,11 +1,11 @@
 import { NextApiHandler } from 'next';
 import { use } from 'next-api-middleware';
 
-import { getAccumulatedVolume } from '@/api/metrics';
+import { getOverview } from '@/api/metrics';
 import { getRequestOnlyMiddleware, logApiErrors } from '@/utils';
 
 const handler: NextApiHandler = async (req, res) => {
-  const data = await getAccumulatedVolume(req.query.TZ as string);
+  const data = await getOverview(req.query.TZ as string);
 
   res.status(200);
   res.send(data);
