--- pages/api/v1/metrics/get-swaps.ts
@@ -1,14 +1,0 @@
-import { NextApiHandler } from 'next';
-import { use } from 'next-api-middleware';
-
-import { getSwaps } from '@/api/metrics';
-import { getRequestOnlyMiddleware, logApiErrors } from '@/utils';
-
-const handler: NextApiHandler = async (req, res) => {
-  const data = await getSwaps(req.query.TZ as string);
-
-  res.status(200);
-  res.send(data);
-};
-
-export default use([getRequestOnlyMiddleware, logApiErrors])(handler);
