+++ views/dapp/v2/metrics/top-tables/top-coins-table/index.tsx
