+++ views/dapp/v2/metrics/top-tables/top-coins-table/top-coins-table-head.tsx
